http://arcemu.org/forums/index.php?showtopic=209&hl= -> "Patch will fix using bad field in a bad way, add support of negative reward_money values as required_money and changing required_money value to rew_money_at_max_level as Walla suggested"

+++ src/ascent-shared/Storage.h
@@ -601,7 +601,7 @@ public:
 
 			case 'i':	// Signed integer
 				*(int32*)&structpointer[offset] = f->GetInt32();
-				offset += sizeof(uint64);
+				offset += sizeof(int32);
 				break;
 
 			case 's':	// Null-terminated string
